TINY-10459: Empty.isEmptyNode unified
