+++ modules/tinymce/src/core/main/ts/delete/MergeBlocks.ts
@@ -5,7 +5,7 @@ import Schema from '../api/html/Schema';
 import * as CaretFinder from '../caret/CaretFinder';
 import CaretPosition from '../caret/CaretPosition';
 import * as ElementType from '../dom/ElementType';
-import * as Empty2 from '../dom/Empty2';
+import * as Empty from '../dom/Empty';
 import * as PaddingBr from '../dom/PaddingBr';
 import * as Parents from '../dom/Parents';
 
@@ -25,11 +25,11 @@ const extractChildren = (block: SugarElement<Element>, schema: Schema): SugarEle
 
 const removeEmptyRoot = (schema: Schema, rootNode: SugarElement<Node>, block: SugarElement<Element>) => {
   const parents = Parents.parentsAndSelf(block, rootNode);
-  return Arr.find(parents.reverse(), (element) => Empty2.isEmpty(schema, element)).each(Remove.remove);
+  return Arr.find(parents.reverse(), (element) => Empty.isEmpty(schema, element)).each(Remove.remove);
 };
 
 const isEmptyBefore = (schema: Schema, el: SugarElement<Node>): boolean =>
-  Arr.filter(Traverse.prevSiblings(el), (el) => !Empty2.isEmpty(schema, el)).length === 0;
+  Arr.filter(Traverse.prevSiblings(el), (el) => !Empty.isEmpty(schema, el)).length === 0;
 
 const nestedBlockMerge = (
   rootNode: SugarElement<Node>,
@@ -38,12 +38,12 @@ const nestedBlockMerge = (
   schema: Schema,
   insertionPoint: SugarElement<Node>
 ): Optional<CaretPosition> => {
-  if (Empty2.isEmpty(schema, toBlock)) {
+  if (Empty.isEmpty(schema, toBlock)) {
     PaddingBr.fillWithPaddingBr(toBlock);
     return CaretFinder.firstPositionIn(toBlock.dom);
   }
 
-  if (isEmptyBefore(schema, insertionPoint) && Empty2.isEmpty(schema, fromBlock)) {
+  if (isEmptyBefore(schema, insertionPoint) && Empty.isEmpty(schema, fromBlock)) {
     Insert.before(insertionPoint, SugarElement.fromTag('br'));
   }
 
@@ -58,8 +58,8 @@ const nestedBlockMerge = (
 const isInline = (schema: Schema, node: SugarElement<Node>): node is SugarElement<HTMLElement> => schema.isInline(SugarNode.name(node));
 
 const sidelongBlockMerge = (rootNode: SugarElement<Node>, fromBlock: SugarElement<Element>, toBlock: SugarElement<Element>, schema: Schema): Optional<CaretPosition> => {
-  if (Empty2.isEmpty(schema, toBlock)) {
-    if (Empty2.isEmpty(schema, fromBlock)) {
+  if (Empty.isEmpty(schema, toBlock)) {
+    if (Empty.isEmpty(schema, fromBlock)) {
       const getInlineToBlockDescendants = (el: SugarElement<Element>) => {
         const helper = (node: SugarElement<Element>, elements: SugarElement<Element>[]): SugarElement<Element>[] =>
           Traverse.firstChild(node).fold(
